# src/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IAUE Student Hub Backend API",
    "version": "1.0.0",
    "description": "Production REST API for IAUE Student Hub backend featuring automated web scraping, academic synchronization, course conflict detection, and administrative review."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server Host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token obtained from /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/announcements": {
      "get": {
        "tags": ["Public"],
        "summary": "List official IAUE announcements with pagination and filtering",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of announcements with IAUE source link" }
        }
      }
    },
    "/api/v1/courses": {
      "get": {
        "tags": ["Public"],
        "summary": "List courses with explicit credit units transparency",
        "parameters": [
          { "name": "level", "in": "query", "schema": { "type": "integer" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of courses" }
        }
      }
    },
...
    "/api/v1/resources": {
      "get": {
        "tags": ["Public"],
        "summary": "List OER materials and official documents",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of academic resources" }
        }
      }
    }
  }
}